requirements.txt
flask
flask-sqlalchemy
docker
flask_jwt_extended